Pathlib used for paths. Reorganised static files.
